Backed out changeset 28f0e71b6292
UltraBlame original commit: 0d6ef98305ad06a00dad18e76de77729c0682b99

+++ accessible/ipc/ProxyAccessible.h
@@ -47,7 +47,7 @@ public:
 
   
   size_t IndexInParent() const { return mParent->mChildren.IndexOf(this); }
-  int32_t IndexOfEmbeddedChild(const ProxyAccessible*);
+  ssize_t IndexOfEmbeddedChild(const ProxyAccessible*);
   bool MustPruneChildren() const;
 
   void Shutdown();
